feat(auth): add ability for the app configure the way it' users to be authenticated
Previously the `base.login` end-point only supports `username` field which is
represent the `name` field in `users` table. In some circumstances we might needs
to be authenticated via `email` field instead.

This way we could configure the way our project is authenticating it' users via
`creasi.base.credentials` config key. The option supports array of strings that
represents the field on `users` table that we want to use as user' credential

Signed-off-by: Fery Wardiyanto <ferywardiyanto@gmail.com>